introduce AvatarDao and RAMAvatarDao
